+++ backend/src/main/java/com/rssmanager/auth/application/AuthService.java
@@ -1,8 +1,8 @@
 package com.rssmanager.auth.application;
 
-import com.rssmanager.auth.controller.dto.CertificateResponse;
-import com.rssmanager.auth.controller.dto.LoginRequest;
-import com.rssmanager.auth.controller.dto.LoginResponse;
+import com.rssmanager.auth.application.dto.CertificateResponse;
+import com.rssmanager.auth.application.dto.LoginRequest;
+import com.rssmanager.auth.application.dto.LoginResponse;
 
 public interface AuthService {
     LoginResponse login(LoginRequest loginRequest);
