+++ backend/src/main/java/com/rssmanager/auth/application/GithubOAuthExchangeService.java
@@ -2,7 +2,7 @@ package com.rssmanager.auth.application;
 
 import com.rssmanager.auth.application.dto.GithubAccessTokenResponse;
 import com.rssmanager.auth.application.dto.GithubUserInfoResponse;
-import com.rssmanager.auth.controller.dto.LoginRequest;
+import com.rssmanager.auth.application.dto.LoginRequest;
 import com.rssmanager.config.GithubOAuthConfig;
 import com.rssmanager.member.domain.Member;
 import java.util.Map;
