+++ backend/src/main/java/com/rssmanager/auth/application/GithubSessionAuthService.java
@@ -1,8 +1,8 @@
 package com.rssmanager.auth.application;
 
-import com.rssmanager.auth.controller.dto.CertificateResponse;
-import com.rssmanager.auth.controller.dto.LoginRequest;
-import com.rssmanager.auth.controller.dto.LoginResponse;
+import com.rssmanager.auth.application.dto.CertificateResponse;
+import com.rssmanager.auth.application.dto.LoginRequest;
+import com.rssmanager.auth.application.dto.LoginResponse;
 import com.rssmanager.member.application.MemberService;
 import com.rssmanager.member.controller.dto.MemberResponse;
 import com.rssmanager.util.SessionManager;
