+++ backend/src/main/java/com/rssmanager/auth/application/OAuthExchangeService.java
@@ -1,6 +1,6 @@
 package com.rssmanager.auth.application;
 
-import com.rssmanager.auth.controller.dto.LoginRequest;
+import com.rssmanager.auth.application.dto.LoginRequest;
 import com.rssmanager.member.domain.Member;
 
 @FunctionalInterface
